Problem 22

+++ src/Problems_20_To_29/problem21.cpp
@@ -11,7 +11,7 @@ using std::cout;
 using std::endl;
 using std::vector;
 
-long sumProperDivisors(int n);
+long sumProperDivisors(long n);
 
 const Problem21 Problem21::INSTANCE = Problem21();
 
@@ -24,8 +24,8 @@ void Problem21::run() const {
             continue;
         }
 
-        int sumOfProperDivisors = sumProperDivisors(i);
-        int sumOfPairsProperDivisors = sumProperDivisors(sumOfProperDivisors);
+        long sumOfProperDivisors = sumProperDivisors(i);
+        long sumOfPairsProperDivisors = sumProperDivisors(sumOfProperDivisors);
 
         if (sumOfPairsProperDivisors == i && i != sumOfProperDivisors) {
             sum += i;
@@ -35,7 +35,7 @@ void Problem21::run() const {
     cout << "The Sum Of Amicable Numbers Below " << max + 1 << " is: " << sum << endl << endl;
 }
 
-long sumProperDivisors(int n) {
+long sumProperDivisors(long n) {
     vector<long> divisors = Factoriser::divisorsFor(n);
 
     long sumOfDivisors = Calculator::sum(divisors);
